// src/shared/model/header.model.ts
export interface MessageCardInterface {
  id: string;
  label: string;
  description: string;
  avatar: string;
  time: string;
}

export interface ContactCardInterface {
  id: string;
  label: string;
  description: string;
  avatar: string;
}

export interface PluginCardInterface {
  id: string;
  label: string;
  description: string;
  avatar: string;
}

export interface KnowledgeBaseCardInterface {
  id: string;
  label: string;
}